Using better hash check command
In addition, added cleanup param

+++ vars/buildAndPushDockerImage.groovy
@@ -11,6 +11,7 @@
  * @param Boolean tagLatest (optional, default=true)
  * @param Boolean forceBuild (optional, default=false)
  * @param String arguments (optional)
+ * @param Boolean cleanup (optional, default=true)
  * @return void
  *
  * @usage examples:
@@ -18,9 +19,10 @@
  *        buildAndPushDockerImage(..., tagLatest:false) //if there's new image to push, it won't tag it as latest
  *        buildAndPushDockerImage(..., forceBuild: true)
  *        buildAndPushDockerImage(..., arguments: "-f MyDockerFile ./my_sub_dir")
+ *        buildAndPushDockerImage(..., arguments: "-f MyDockerFile ./my_sub_dir", cleanup: false)
  */
 
-def call(String registry, String creds, String image, String tag, Boolean tagLatest=true, Boolean forceBuild=false, String arguments='') {
+def call(String registry, String creds, String image, String tag, Boolean tagLatest=true, Boolean forceBuild=false, String arguments='', Boolean cleanup=true) {
   for (var in [registry, creds, image, tag]) {
     if (var == null || var == '') {
       throw new Exception("Method 'buildAndPushDockerImage' must contain params 'registry' ,'creds', 'image', 'tag' !!!")
@@ -35,17 +37,17 @@ def call(String registry, String creds, String image, String tag, Boolean tagLat
     arguments = "."
   }
   
-  sh "docker pull ${dockerImageLatest} || true"
   echo "Building docker image ${dockerImageFull}"
   docker.withRegistry("${registry}", "${dockerCredential}") {
+    sh "docker pull ${dockerImageLatest} || true"
     def newImage
     if (!forceBuild) {
       newImage = docker.build(dockerImageFull, "--cache-from ${dockerImageLatest} ${arguments}")
-      image_hash = sh(returnStdout: true, script: "docker inspect --format='{{index .RepoDigests 0}}' ${dockerImageFull}").trim()
+      image_hash = sh(returnStdout: true, script: "docker images --no-trunc --quiet ${dockerImageFull}").trim()
       def latest_hash = ''
       latest_exist = sh(returnStdout: true, script: "docker images -q ${dockerImageLatest} | wc -l").trim() as Integer
       if (latest_exist > 0 ) {
-        latest_hash = sh(returnStdout: true, script: "docker inspect --format='{{index .RepoDigests 0}}' ${dockerImageLatest}").trim()
+        latest_hash = sh(returnStdout: true, script: "docker images --no-trunc --quiet ${dockerImageLatest}").trim()
       }
       if (image_hash != latest_hash) {
         echo "Pushing new image"
@@ -69,14 +71,17 @@ def call(String registry, String creds, String image, String tag, Boolean tagLat
   }
   
   //Cleanup
-  sh "docker rmi ${dockerImageFull}"
-  sh "docker rmi ${dockerImageLatest} || true"
-  sh "docker images -f dangling=true | grep ${dockerRegistry}/${dockerImageName} | awk '{print \$3}' | xargs --no-run-if-empty docker rmi"
+  if (cleanup) {
+    sh "docker rmi ${dockerImageFull}"
+    sh "docker rmi ${dockerImageLatest} || true"
+    sh "docker images -f dangling=true | grep ${dockerRegistry}/${dockerImageName} | awk '{print \$3}' | xargs --no-run-if-empty docker rmi"
+  }
 }
 
 def call(Map config) {
   if (!config.containsKey('tagLatest')) { config.tagLatest = true }
   if (!config.containsKey('forceBuild')) { config.forceBuild = false }
   if (!config.containsKey('arguments')) { config.arguments = '' }
-  call(config.registry, config.creds, config.image, config.tag, config.tagLatest, config.forceBuild, config.arguments)
+  if (!config.containsKey('cleanup')) { config.cleanup = true }
+  call(config.registry, config.creds, config.image, config.tag, config.tagLatest, config.forceBuild, config.arguments, config.cleanup)
 }
